backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
rembg[cpu]==2.0.58  # Latest version with session improvements

# Alternative background removal libraries (Phase 0 research)
torch==2.1.1
torchvision==0.16.1
onnxruntime==1.16.3

# AWS and serverless
boto3==1.34.0
mangum==0.17.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
redis==5.0.1

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4